server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0,<1.0.0",
    "bcrypt>=4.2.0,<5.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.0,<1.0.0",
    "pipecat-ai[cartesia,livekit,openai,soniox]>=0.0.108",
    "pyjwt[crypto]>=2.10.0,<3.0.0",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.42.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]
